+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.11.0
+## unreleased
 
 * move tmpFolder into local Project to avoid conflicts with limited space in /tmp folder